Resolve gradient issue for android 6
Handle upgrade scenario
