+++ android/java/org/chromium/chrome/browser/DormantUsersEngagementDialogFragment.java
@@ -23,6 +23,7 @@ import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import androidx.appcompat.widget.AppCompatImageView;
 import androidx.fragment.app.DialogFragment;
 
 import org.chromium.base.ContextUtils;
