+++ android/java/org/chromium/chrome/browser/app/BraveActivity.java
@@ -555,60 +555,20 @@ public abstract class BraveActivity<C extends ChromeActivityComponent> extends C
                         BraveVpnUtils.SUBSCRIPTION_PARAM_TEXT, getPackageName());
             }
         }
-        if (OnboardingPrefManager.getInstance().isDormantUsersEngagementEnabled()
-                || getPackageName().equals(BRAVE_PRODUCTION_PACKAGE_NAME)) {
-            OnboardingPrefManager.getInstance().setDormantUsersNotificationTime(
-                    RetentionNotificationUtil.DORMANT_USERS_DAY_14, setTimeInMillis(14 * 24 * 60));
-            OnboardingPrefManager.getInstance().setDormantUsersNotificationTime(
-                    RetentionNotificationUtil.DORMANT_USERS_DAY_25, setTimeInMillis(25 * 24 * 60));
-            OnboardingPrefManager.getInstance().setDormantUsersNotificationTime(
-                    RetentionNotificationUtil.DORMANT_USERS_DAY_40, setTimeInMillis(40 * 24 * 60));
+        if (PackageUtils.isFirstInstall(this)
+                && (OnboardingPrefManager.getInstance().isDormantUsersEngagementEnabled()
+                        || getPackageName().equals(BRAVE_PRODUCTION_PACKAGE_NAME))) {
+            OnboardingPrefManager.getInstance().setDormantUsersPrefs();
             if (!OnboardingPrefManager.getInstance().isDormantUsersNotificationsStarted()) {
-                RetentionNotificationUtil.scheduleNotificationWithTime(this,
-                        RetentionNotificationUtil.DORMANT_USERS_DAY_14,
-                        OnboardingPrefManager.getInstance().getDormantUsersNotificationTime(
-                                RetentionNotificationUtil.DORMANT_USERS_DAY_14));
-                RetentionNotificationUtil.scheduleNotificationWithTime(this,
-                        RetentionNotificationUtil.DORMANT_USERS_DAY_25,
-                        OnboardingPrefManager.getInstance().getDormantUsersNotificationTime(
-                                RetentionNotificationUtil.DORMANT_USERS_DAY_25));
-                RetentionNotificationUtil.scheduleNotificationWithTime(this,
-                        RetentionNotificationUtil.DORMANT_USERS_DAY_40,
-                        OnboardingPrefManager.getInstance().getDormantUsersNotificationTime(
-                                RetentionNotificationUtil.DORMANT_USERS_DAY_40));
+                RetentionNotificationUtil.scheduleDormantUsersNotifications(this);
                 OnboardingPrefManager.getInstance().setDormantUsersNotificationsStarted(true);
             }
         }
     }
 
-    private long setTimeInMillis(int timeInMinutes) {
-        Calendar calendar = Calendar.getInstance();
-        calendar.setTimeInMillis(System.currentTimeMillis());
-        calendar.add(Calendar.MINUTE, timeInMinutes);
-
-        Date date = calendar.getTime();
-        return date.getTime();
-    }
-
-    private void setDormantUsersPrefs() {
-        OnboardingPrefManager.getInstance().setDormantUsersNotificationTime(
-                RetentionNotificationUtil.DORMANT_USERS_DAY_14, setTimeInMillis(14 * 24 * 60));
-        OnboardingPrefManager.getInstance().setDormantUsersNotificationTime(
-                RetentionNotificationUtil.DORMANT_USERS_DAY_25, setTimeInMillis(25 * 24 * 60));
-        OnboardingPrefManager.getInstance().setDormantUsersNotificationTime(
-                RetentionNotificationUtil.DORMANT_USERS_DAY_40, setTimeInMillis(40 * 24 * 60));
-        RetentionNotificationUtil.scheduleNotificationWithTime(this,
-                RetentionNotificationUtil.DORMANT_USERS_DAY_14,
-                OnboardingPrefManager.getInstance().getDormantUsersNotificationTime(
-                        RetentionNotificationUtil.DORMANT_USERS_DAY_14));
-        RetentionNotificationUtil.scheduleNotificationWithTime(this,
-                RetentionNotificationUtil.DORMANT_USERS_DAY_25,
-                OnboardingPrefManager.getInstance().getDormantUsersNotificationTime(
-                        RetentionNotificationUtil.DORMANT_USERS_DAY_25));
-        RetentionNotificationUtil.scheduleNotificationWithTime(this,
-                RetentionNotificationUtil.DORMANT_USERS_DAY_40,
-                OnboardingPrefManager.getInstance().getDormantUsersNotificationTime(
-                        RetentionNotificationUtil.DORMANT_USERS_DAY_40));
+    public void setDormantUsersPrefs() {
+        OnboardingPrefManager.getInstance().setDormantUsersPrefs();
+        RetentionNotificationUtil.scheduleDormantUsersNotifications(this);
     }
 
     private final ConnectivityManager.NetworkCallback mNetworkCallback =
