+++ android/java/org/chromium/chrome/browser/upgrade/BravePackageReplacedBroadcastReceiver.java
@@ -9,6 +9,9 @@ import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 
+import org.chromium.chrome.browser.app.BraveActivity;
+import org.chromium.chrome.browser.notifications.retention.RetentionNotificationUtil;
+import org.chromium.chrome.browser.onboarding.OnboardingPrefManager;
 import org.chromium.chrome.browser.preferences.BravePreferenceKeys;
 import org.chromium.chrome.browser.preferences.SharedPreferencesManager;
 
@@ -25,6 +28,16 @@ public final class BravePackageReplacedBroadcastReceiver extends BroadcastReceiv
         BraveUpgradeJobIntentServiceImpl.maybePerformUpgradeTasks(context);
         try {
             SharedPreferencesManager.getInstance().writeInt(BravePreferenceKeys.BRAVE_APP_OPEN_COUNT, 0);
+            // Set dormant users notifications
+            if (OnboardingPrefManager.getInstance().isDormantUsersEngagementEnabled()
+                    || context.getPackageName().equals(
+                            BraveActivity.BRAVE_PRODUCTION_PACKAGE_NAME)) {
+                OnboardingPrefManager.getInstance().setDormantUsersPrefs();
+                if (!OnboardingPrefManager.getInstance().isDormantUsersNotificationsStarted()) {
+                    RetentionNotificationUtil.scheduleDormantUsersNotifications(context);
+                    OnboardingPrefManager.getInstance().setDormantUsersNotificationsStarted(true);
+                }
+            }
         } catch (Exception exc) {
             // Sometimes the pref is not registered yet in the app
         }
