+++ android/java/res/layout/fragment_dormant_users_engagement_dialog.xml
@@ -22,12 +22,11 @@
         android:textColor="@color/shield_text_color"
         android:textStyle="bold" />
 
-    <ImageView
+    <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/image_view"
         android:layout_width="200dp"
         android:layout_height="175dp"
         android:layout_marginBottom="16dp"
-        android:src="@drawable/ic_brave_mobiledata"
         android:layout_gravity="center_horizontal"
         android:contentDescription="@null"/>
